controllers rename y reubicacion carpeta imagenes
