--- Integrador Final/demo/src/main/java/com/example/controles/PrimaryController.java
@@ -1,15 +1,0 @@
-package com.example.controles;
-
-import java.io.IOException;
-
-import com.example.App;
-
-import javafx.fxml.FXML;
-
-public class PrimaryController {
-
-    @FXML
-    private void switchToSecondary() throws IOException {
-        App.setRoot("secondary");
-    }
-}
