--- Integrador Final/demo/src/main/java/com/example/controles/SecondaryController.java
@@ -1,15 +1,0 @@
-package com.example.controles;
-
-import java.io.IOException;
-
-import com.example.App;
-
-import javafx.fxml.FXML;
-
-public class SecondaryController {
-
-    @FXML
-    private void switchToPrimary() throws IOException {
-        App.setRoot("primary");
-    }
-}
