+++ Integrador Final/demo/src/main/java/com/example/controllers/PrimaryController.java
@@ -1,0 +1,15 @@
+package com.example.controllers;
+
+import java.io.IOException;
+
+import com.example.App;
+
+import javafx.fxml.FXML;
+
+public class PrimaryController {
+
+    @FXML
+    private void switchToSecondary() throws IOException {
+        App.setRoot("secondary");
+    }
+}
