+++ Integrador Final/demo/src/main/java/com/example/controllers/RegistroController.java
@@ -1,4 +1,4 @@
-package com.example.controles;
+package com.example.controllers;
 
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
