+++ Integrador Final/demo/src/main/java/com/example/controllers/SecondaryController.java
@@ -1,0 +1,15 @@
+package com.example.controllers;
+
+import java.io.IOException;
+
+import com.example.App;
+
+import javafx.fxml.FXML;
+
+public class SecondaryController {
+
+    @FXML
+    private void switchToPrimary() throws IOException {
+        App.setRoot("primary");
+    }
+}
