+++ Integrador Final/demo/src/main/resources/Imagenes/Registro.png
